Add request validation by JSON Schema as config switch to CLI_v1
