+++ integration/config/imageREST_v1.schema
@@ -1,0 +1,1 @@
+../../python/lsst/dax/imgserv/config/imageREST_v1.schema
