+++ integration/config/settings.json
@@ -1,5 +1,6 @@
 {
   "DEBUG": true,
+  "DAX_IMG_VALIDATE": true,
   "DAX_IMG_DBCONF": "~/.lsst/dbAuth-dbServ.ini",
   "DAX_IMG_DS": "./datasets/sdss/preprocessed/dr7/sdss_stripe82_00",  
   "DAX_IMG_DB": "DC_W13_Stripe82",
