+++ integration/input_v1/test_calexp_i10_v1.json
@@ -1,33 +1,31 @@
 {
   "name": "calexp_i10_v1",
-  "get": {
-    "image": {
-      "db": "DC_W13_Stripe82",
-      "ds": "calexp",
-      "cutout": {
-        "image_id": {
-          "nearest": { 
-            "center": { 
-              "x": 37.6444598, 
-              "y": 0.104625,
-              "unit": "deg"
-            },
-            "filter": "r"
-          }
-        },
-        "center": {
-          "x": 37.6444598,
-          "y": 0.104625,
-          "unit": "deg"
-        },
-        "size": {
-          "x": 100,
-          "y": 100,
-          "unit": "pixel"
-          }
+  "image": {
+    "db": "DC_W13_Stripe82",
+    "ds": "calexp",
+    "cutout": {
+      "image_id": {
+        "nearest": { 
+          "center": { 
+            "x": 37.6444598, 
+            "y": 0.104625,
+            "unit": "deg"
+          },
+          "filter": "r"
         }
-    },
-    "api_id": [ "db", "ds", "nearest.filter", "center.x", "center.y", "center.unit", "size.x", "size.y", "size.unit" ]
+      },
+      "center": {
+        "x": 37.6444598,
+        "y": 0.104625,
+        "unit": "deg"
+      },
+      "size": {
+        "x": 100,
+        "y": 100,
+        "unit": "pixel"
+        }
+      }
   },
-  "check": { "sha256": "a3bac7226f57ef116ee76bd0e0e33e1da1a6809e9d65d09c9b5ce12f54c4fefe" }
+  "api_id": [ "db", "ds", "filter", "center.x", "center.y", "center.unit", "size.x", "size.y", "size.unit" ],
+  "check": { "sha256": "5875b82425c7b749a8e522662a0ddc11f90b5c20c01e8998a934760e272eb27b" }
 }
