+++ integration/input_v1/test_calexp_i6_v1.json
@@ -1,21 +1,19 @@
 {
   "name": "calexp_i6_v1",
-  "get": {
-    "image": {
-      "db": "DC_W13_Stripe82",
-      "ds": "calexp",
-      "image_id": {
-        "nearest": {
-          "center": { 
-            "x": 37.644598, 
-            "y": 0.104625,
-            "unit": "deg"
-          },
-          "filter": "r"
-        }
+  "image": {
+    "db": "DC_W13_Stripe82",
+    "ds": "calexp",
+    "image_id": {
+      "nearest": {
+        "center": { 
+          "x": 37.644598, 
+          "y": 0.104625,
+          "unit": "deg"
+        },
+        "filter": "r"
       }
-    },
-    "api_id": [ "db", "ds", "nearest.filter", "center.x", "center.y", "center.unit" ]
+    }
   },
-  "check": { "sha256": "fe1675878b720a5bac33484b211bd8ff2e3cf2892a4ea4aa855b07c435511451" }
+  "api_id": [ "db", "ds", "filter", "center.x", "center.y", "center.unit" ],
+  "check": { "sha256": "ded6b991ea987174135e9f953b6328ff4ca6549a85841c8b6ed36d003e74b63d" }
 }
