+++ integration/input_v1/test_calexp_i7_v1.json
@@ -1,21 +1,19 @@
 {
   "name": "calexp_i7_v1",
-  "get": {
-    "image": {
-      "db": "DC_W13_Stripe82",
-      "ds": "calexp",
-      "image_id": {
-        "data_id": {
-          "rcff": {
-            "run": 5646, 
-            "camcol": 4, 
-            "field": 694, 
-            "filter": "g" 
-          }
+  "image": {
+    "db": "DC_W13_Stripe82",
+    "ds": "calexp",
+    "image_id": {
+      "data_id": {
+        "rcff": {
+          "run": 5646, 
+          "camcol": 4, 
+          "field": 694, 
+          "filter": "g" 
         }
       }
-    },
-    "api_id": [ "db", "ds", "run", "camcol", "field", "filter" ]
+    }
   },
+  "api_id": [ "db", "ds", "run", "camcol", "field", "filter" ],
   "check": { "sha256": "ded6b991ea987174135e9f953b6328ff4ca6549a85841c8b6ed36d003e74b63d" }
 }
