+++ integration/input_v1/test_coadd_trim_v1.json
@@ -1,31 +1,29 @@
 {
   "name": "coadd_trim_v1",
-  "get": {
-    "image": {
-      "db": "DC_W13_Stripe82",
-      "ds": "deepcoadd",
-      "image_id": {
-        "data_id": {
-          "tpf": {
-            "tract": 0, 
-            "patch_x": 321, 
-            "patch_y": 4, 
-            "filter": "r" 
-          }
-        },
-        "center": {
-          "x": 19.36995,
-          "y": -0.3146,
-          "unit": "deg" 
-        },
-        "size": {
-          "x": 300,
-          "y": 300,
-          "unit": "pixel" 
+  "image": {
+    "db": "DC_W13_Stripe82",
+    "ds": "deepcoadd",
+    "image_id": {
+      "data_id": {
+        "tpf": {
+          "tract": 0, 
+          "patch_x": 321, 
+          "patch_y": 4, 
+          "filter": "r" 
         }
+      },
+      "center": {
+        "x": 19.36995,
+        "y": -0.3146,
+        "unit": "deg" 
+      },
+      "size": {
+        "x": 300,
+        "y": 300,
+        "unit": "pixel" 
       }
-    },
-    "api_id": [ "db", "ds", "tract", "patch_x", "patch_y", "filter", "center.x", "center.y", "center.unit", "size.x", "size.y", "size.unit" ] 
+    }
   },
+  "api_id": [ "db", "ds", "tract", "patch_x", "patch_y", "filter", "center.x", "center.y", "center.unit", "size.x", "size.y", "size.unit" ],
   "check": { "sha256": "f7d40802282a15b09c910d1964e281fa4a51bb7a94934ee389267920ee8b5f26" }
 }
