+++ integration/input_v1/test_deepcoadd_i13_v1.json
@@ -1,21 +1,19 @@
 {
   "name": "deepcoadd_i13_v1",
-  "get": {
-    "image": {
-      "db": "DC_W13_Stripe82",
-      "ds": "deepcoadd",
-      "image_id": {
-          "nearest": {
-            "center": { 
-              "x": 19.36995, 
-              "y": -0.3146,
-              "unit": "deg"
-            },
-            "filter": "r"
-        }
+  "image": {
+    "db": "DC_W13_Stripe82",
+    "ds": "deepcoadd",
+    "image_id": {
+        "nearest": {
+          "center": { 
+            "x": 19.36995, 
+            "y": -0.3146,
+            "unit": "deg"
+          },
+          "filter": "r"
       }
-    },
-    "api_id": [ "db", "ds", "nearest.filter", "center.x", "center.y", "center.unit" ] 
+    }
   },
+  "api_id": [ "db", "ds", "filter", "center.x", "center.y", "center.unit" ],
   "check": { "sha256": "f7d40802282a15b09c910d1964e281fa4a51bb7a94934ee389267920ee8b5f26" }
 }
