+++ integration/input_v1/test_deepcoadd_i14_v1.json
@@ -1,21 +1,19 @@
 {
   "name": "deepcoadd_i14_v1",
-  "get": {
-    "image": {
-      "db": "DC_W13_Stripe82",
-      "ds": "deepcoadd",
-      "image_id": {
-        "data_id": {
-          "tpf": {
-            "tract": 0, 
-            "patch_x": 225, 
-            "patch_y": 1, 
-            "filter": "i" 
-          }
+  "image": {
+    "db": "DC_W13_Stripe82",
+    "ds": "deepcoadd",
+    "image_id": {
+      "data_id": {
+        "tpf": {
+          "tract": 0, 
+          "patch_x": 225, 
+          "patch_y": 1, 
+          "filter": "i" 
         }
       }
-    },
-    "api_id": [ "db", "ds", "tract", "patch_x", "patch_y", "filter" ] 
+    }
   },
+  "api_id": [ "db", "ds", "tract", "patch_x", "patch_y", "filter" ],
   "check": { "sha256": "21a188c1c5f6177f9d74577a0f6eef353fc67570f89ea4a612aabf2de263b5ae" }
 }
