+++ integration/input_v1/test_deepcoadd_i15_v1.json
@@ -1,14 +1,12 @@
 {
   "name": "deepcoadd_i15_v1",
-  "get": {
-    "image": {
-      "db": "DC_W13_Stripe82",
-      "ds": "deepcoadd",
-      "image_id": {
-        "science_id": 23986176
-      }
-    },
-    "api_id": [ "db", "ds", "science_id" ] 
+  "image": {
+    "db": "DC_W13_Stripe82",
+    "ds": "deepcoadd",
+    "image_id": {
+      "science_id": 23986176
+    }
   },
+  "api_id": [ "db", "ds", "science_id" ],
   "check": { "sha256": "99e20c0127f0145ac89240be326fcfb6232458ef479351b218e8d7918e57f5f3" }
 }
