+++ integration/input_v1/test_deepcoadd_i16_v1.json
@@ -1,33 +1,31 @@
 {
   "name": "deepcoadd_i16_v1",
-  "get": {
-    "image": {
-      "db": "DC_W13_Stripe82",
-      "ds": "deepcoadd",
-      "cutout": {
-        "image_id": {
-          "nearest": { 
-            "center": { 
-              "x": 19.36995, 
-              "y": -0.3146,
-              "unit": "deg"
-            },
-            "filter": "r"
-          }
-        },
-        "center": {
-          "x": 19.36995,
-          "y": -0.3146,
-          "unit": "deg"
-        },
-        "size": {
-          "x": 115,
-          "y": 235,
-          "unit": "arcsec"
+  "image": {
+    "db": "DC_W13_Stripe82",
+    "ds": "deepcoadd",
+    "cutout": {
+      "image_id": {
+        "nearest": { 
+          "center": { 
+            "x": 19.36995, 
+            "y": -0.3146,
+            "unit": "deg"
+          },
+          "filter": "r"
         }
+      },
+      "center": {
+        "x": 19.36995,
+        "y": -0.3146,
+        "unit": "deg"
+      },
+      "size": {
+        "x": 115,
+        "y": 235,
+        "unit": "arcsec"
       }
-    },
-    "api_id": [ "db", "ds", "nearest.filter", "center.x", "center.y", "center.unit", "size.x", "size.y", "size.unit" ]
+    }
   },
+  "api_id": [ "db", "ds", "filter", "center.x", "center.y", "center.unit", "size.x", "size.y", "size.unit" ],
   "check": { "sha256": "70d9a849c2d1b7e870c802f585439a05c8deb0cf7e6425f65078689cfb67c5d4" }
 }
