+++ integration/input_v1/test_deepcoadd_i17_v1.json
@@ -1,33 +1,31 @@
 {
   "name": "deepcoadd_i17_v1",
-  "get": {
-    "image": {
-      "db": "DC_W13_Stripe82",
-      "ds": "deepcoadd",
-      "cutout": {
-        "image_id": {
-          "nearest": { 
-            "center": {
-              "x": 19.62322, 
-              "y": -0.2978,
-              "unit": "deg"
-            },
-            "filter": "r"
-          },  
-          "center": { 
+  "image": {
+    "db": "DC_W13_Stripe82",
+    "ds": "deepcoadd",
+    "cutout": {
+      "image_id": {
+        "nearest": { 
+          "center": {
             "x": 19.62322, 
             "y": -0.2978,
             "unit": "deg"
           },
-          "size": {
-            "x": 100,
-            "y": 100,
-            "unit": "pixel"
-          }
+          "filter": "r"
+        },  
+        "center": { 
+          "x": 19.62322, 
+          "y": -0.2978,
+          "unit": "deg"
+        },
+        "size": {
+          "x": 100,
+          "y": 100,
+          "unit": "pixel"
         }
       }
-    },
-    "api_id": [ "db", "ds", "nearest.filter", "center.x", "center.y", "center.unit", "size.x", "size.y", "size.unit" ]
+    }
   },
+  "api_id": [ "db", "ds", "filter", "center.x", "center.y", "center.unit", "size.x", "size.y", "size.unit" ],
   "check": { "sha256": "aa926d99e488cc79944a0433f97ead8abc222f818bbedc79863ff4bc19d697e2" }
 }
