+++ integration/input_v1/test_deepcoadd_i20_v1.json
@@ -1,26 +1,24 @@
 {
   "name": "deepcoadd_i20_v1",
-  "get": {
-    "image": {
-      "db": "DC_W13_Stripe82",
-      "ds": "deepcoadd",
-      "cutout": {
-        "image_id": {
-          "science_id": 23986176
-        },
-        "center": { 
-          "x": 9.47103, 
-          "y": -1.181766,
-          "unit": "deg"
-        },
-        "size": {
-          "x": 30.0,
-          "y": 45.0,
-          "unit": "arcsec"
-        }
+  "image": {
+    "db": "DC_W13_Stripe82",
+    "ds": "deepcoadd",
+    "cutout": {
+      "image_id": {
+        "science_id": 23986176
+      },
+      "center": { 
+        "x": 9.47103, 
+        "y": -1.181766,
+        "unit": "deg"
+      },
+      "size": {
+        "x": 30.0,
+        "y": 45.0,
+        "unit": "arcsec"
       }
-    },
-    "api_id": [ "db", "ds", "science_id", "center.x", "center.y", "center.unit", "size.x", "size.y", "size.unit" ]
+    }
   },
+  "api_id": [ "db", "ds", "science_id", "center.x", "center.y", "center.unit", "size.x", "size.y", "size.unit" ],
   "check": { "sha256": "245add981b77b62debd6a2882c51987b4ecb1d138deadd35a6ca084567fc9b77" }
 }
