+++ integration/input_v1/test_raw_i2_v1.json
@@ -1,21 +1,19 @@
 {
   "name": "raw_i2_v1",
-  "get": {
-    "image": {
-      "db": "DC_W13_Stripe82",
-      "ds": "raw",
-      "image_id": {
-          "data_id": { 
-            "rcff": { 
-              "run": 5646, 
-              "camcol": 4, 
-              "field": 694, 
-              "filter": "g" 
-            }
+  "image": {
+    "db": "DC_W13_Stripe82",
+    "ds": "raw",
+    "image_id": {
+        "data_id": { 
+          "rcff": { 
+            "run": 5646, 
+            "camcol": 4, 
+            "field": 694, 
+            "filter": "g" 
           }
         }
-      },
-    "api_id": [ "db", "ds", "run", "camcol", "field", "filter" ]
-  },
+      }
+    },
+  "api_id": [ "db", "ds", "run", "camcol", "field", "filter" ],
   "check": { "sha256": "0e23206bd029c6b225356f1f88b6044136e51718b7e1afa6f4fec677a97c0b47" }
 }
