+++ python/lsst/dax/imgserv/config/api_map.json
@@ -4,10 +4,10 @@
     "Source: https://lsst.org/schemas/imgserv_api_v1.schema",
     "Note: this file was generated by hand, but can be automated."
   ],
-  "416e52da3590b583cf3e237849c15ebf9d73b203eaa9ba8d36bf20a07eba6362": { 
+  "2ca69ed8aa69c9f01469617000a6a30d5df02d3843677330c6b457eef0e5b0d9": { 
     "api": "Image.full_nearest",
     "hash_op": "hashlib.sha256(str(sorted(api_id)).encode('utf-8')).hexdigest()",
-    "api_id": ["db", "ds", "nearest.filter", "center.x", "center.y", "center.unit"]
+    "api_id": ["db", "ds", "filter", "center.x", "center.y", "center.unit"]
   },
   "1697b5040cb4fb249ab5152412cc95a3b7f5a9d74dd5ed5666cacb1e420fb539": {
     "api": "Image.full_from_data_id",
@@ -29,10 +29,10 @@
     "hash_op": "hashlib.sha256(str(sorted(api_id)).encode('utf-8')).hexdigest()",
     "api_id": ["db", "ds", "science_id"]
   },
-  "fa4e7e1117f48820092661c0a1534c655cd9c819bd01c00941d4d85c1e9e8951": {
+  "415d9f7798aff2a6a60e6d0835d1d470119e057f7a3164b5cc1148cfba2c467a": {
     "api": "Image.cutout_from_nearest",
     "hash_op": "hashlib.sha256(str(sorted(api_id)).encode('utf-8')).hexdigest()",
-    "api_id": ["db", "ds", "nearest.filter", "center.x", "center.y", "center.unit", "size.x", "size.y", "size.unit"]
+    "api_id": ["db", "ds", "filter", "center.x", "center.y", "center.unit", "size.x", "size.y", "size.unit"]
   },
   "d27cb5beabb1aa858c6d7f29a744a44b33477d72e14068f2ac66749b1b1bb348": {
     "api": "Image.cutout_from_data_id",
@@ -53,6 +53,11 @@
     "api": "Image.cutout_from_science_id",
     "hash_op": "hashlib.sha256(str(sorted(api_id)).encode('utf-8')).hexdigest()",
     "api_id": ["db", "ds", "science_id", "center.x", "center.y", "center.unit", "size.x", "size.y", "size.unit"]
+  },
+  "b7f532768eaa35770d49825c046839565db3602381d1d82f7360634b28c45ae4": {
+    "api": "Image.cutout_from_skymap_id",
+    "hash_op": "hashlib.sha256(str(sorted(api_id)).encode('utf-8')).hexdigest()",
+    "api_id": ["db", "ds", "skymap_id", "filter", "center.x", "center.y", "center.unit", "size.x", "size.y", "size.unit"]
   }
 }
