+++ ups/dax_imgserv.table
@@ -12,6 +12,7 @@ setupRequired(scons)
 setupRequired(sconsUtils)
 setupRequired(skymap)
 setupRequired(sqlalchemy)
+setupRequired(jsonschema)
 
 envPrepend(LD_LIBRARY_PATH, ${PRODUCT_DIR}/lib)
 envPrepend(DYLD_LIBRARY_PATH, ${PRODUCT_DIR}/lib)
